refactor: extract service config
